+++ src/ExplicitClasslistAnnotationDriver.php
@@ -9,10 +9,19 @@ use Doctrine\ORM\Mapping\Driver\AnnotationDriver;
 
 class ExplicitClasslistAnnotationDriver extends AnnotationDriver
 {
-    public function __construct(Reader $reader, array $entity_classes)
+    /**
+     * ExplicitClasslistAnnotationDriver constructor.
+     *
+     * @param Reader     $reader
+     * @param array|NULL $entity_classes The list of entity class names
+     *                                   [NB] it is not expected to be valid for this to be empty at runtime, but
+     *                                   allowing a null value allows us to create an instance in development / test
+     *                                   environments without full config.
+     */
+    public function __construct(Reader $reader, array $entity_classes = NULL)
     {
         parent::__construct($reader, []);
-        $this->classNames = $entity_classes;
+        $this->classNames = $entity_classes ?: [];
     }
 
     public function getAllClassNames()